+++ mods/src/multiplayer/inventory.cpp
@@ -1,0 +1,10 @@
+// send inventory every 10 ticks or when leaving game
+
+// allow receiving armor data from server
+
+// use negated entity ID so items don't drop
+// - make negation a global function?
+
+// Add SetSpawnPositionPacket to syncing.cpp under "Sync Level Spawn Position"
+
+// Only save player-data if inventory dats is present